+++ ember-aria-utilities/src/modifiers/-private/navigation.ts
@@ -14,18 +14,10 @@ import {
   upCell,
 } from './focus';
 import { DOWN, END, HOME, LEFT, PAGEDOWN, PAGEUP, RIGHT, UP } from './keys';
+import { isNavDisabled } from './navigation-state';
 import { isCell, nonHeaderCell } from './node-selectors';
 
 export const NAV_KEYS = [LEFT, DOWN, UP, RIGHT, HOME, END, PAGEDOWN, PAGEUP];
-
-// presence means disabled
-const disabledGrids = new WeakSet<Element>();
-
-export const toggleNav = (grid: Element) =>
-  isNavDisabled(grid) ? enableNav(grid) : disableNav(grid);
-export const isNavDisabled = (grid: Element) => disabledGrids.has(grid);
-export const disableNav = (grid: Element) => disabledGrids.add(grid);
-export const enableNav = (grid: Element) => disabledGrids.delete(grid);
 
 export function handleNavigation(
   event: KeyboardEvent,
